split resource controller and route
