+++ run.py
@@ -1,7 +1,7 @@
 from tutor import app, db
 from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
-from tutor.routes import home, course
+from tutor.routes import home, course, resources
 from tutor.routes import model
